refactor : 폴더 이름을 complaint-apply로 변경한다
