+++ src/Router.tsx
@@ -7,10 +7,10 @@ import Detail from "./pages/detail/Detail";
 import View from "./pages/view/View";
 import MyPage from "./pages/my-page/MyPage";
 import Search from "./pages/search/Search";
-import AcademicInfoStep from "./pages/application-step/AcademicInfoStep";
-import CategorySelectionStep from "./pages/application-step/CategorySelectionStep";
-import CompletedStep from "./pages/application-step/CompletedStep";
-import ComplaintsWrittingStep from "./pages/application-step/CompaintsWrittingStep";
+import AcademicInfoStep from "./pages/complaint-apply/AcademicInfoStep";
+import CategorySelectionStep from "./pages/complaint-apply/CategorySelectionStep";
+import CompletedStep from "./pages/complaint-apply/CompletedStep";
+import ComplaintsWrittingStep from "./pages/complaint-apply/CompaintsWrittingStep";
 import Login from "./pages/login/Login";
 import Redirect from "./pages/login/Redirect";
 import PrivateRoute from "./components/PrivateRoute";
